refactor: validateCarName 모듈화
