+++ src/App.js
@@ -1,22 +1,6 @@
 import { Console } from '@woowacourse/mission-utils';
+import { validateCarName } from './validation/index.js';
 
-const validateCarName = (input) => {
-  const regex = /^[A-Za-z0-9]+$/;
-  const numberName = Number(input);
-  if (input.length > 5) {
-    throw new Error('[ERROR] 자동차 이름은 5자 이하로 입력해주세요.');
-  }
-
-  if (!regex.test(input)) {
-    throw new Error(
-      '[ERROR] 자동차 이름에는 공백을 제외한 영어, 숫자만 사용할 수 있어요.'
-    );
-  }
-
-  if (numberName === 0) {
-    throw new Error('[ERROR] 자동차 이름에는 공백을 입력할 수 없어요.');
-  }
-};
 const determineCarName = (input) => {
   const names = input.split(',');
   names.forEach((name) => validateCarName(name));
